requirements.txt
pillow
pandas
torch
torchvision
